Extract main component
